pipeline/functions/ParagraphExtraction/requirements.txt
alembic==1.4.2
azure-core==1.5.0
azure-storage-blob==12.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
configparser==5.0.0
cryptography==2.9.2
cycler==0.10.0
databricks-cli==0.10.0
docker==4.2.0
elasticsearch==7.7.0
entrypoints==0.3
es-pandas==0.0.10
espandas==1.0.4
et-xmlfile==1.0.1
Flask==1.1.2
fuzzywuzzy==0.18.0
gitdb==4.0.5
GitPython==3.1.2
gorilla==0.3.0
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.2.0
lxml==4.5.1
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mlflow==1.8.0
msrest==0.6.14
numpy==1.18.4
oauthlib==3.1.0
openpyxl==3.0.3
pandas==1.0.3
pandasticsearch==0.5.3
Pillow==7.1.2
progressbar2==3.51.3
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
protobuf==3.12.1
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
python-Levenshtein==0.12.0
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
xlrd==1.2.0
